Moved the navigation component to organisms category
